New ranking system, less API calls, reversed processing.
Search results are now all taken into account to extract the best according to a weighted score. The "top_hit" provided by Tidal was too often inaccurate.
When a track is already in your favorites, the program moves on to the next one and no API call is made.
Sounds are now processed in reverse order for greater chronological consistency with YTMusic.